requirements.txt
kafka-python==2.0.2
psycopg2==2.8.6
PyYAML==5.4.1
requests==2.25.1
